Change io_flags type

+++ src/num/ops_opts.c
@@ -24,7 +24,7 @@ void print_operator_run_flags(int N, operator_run_opt_flags_t run_opts[N][N])
 		debug_printf(DP_INFO, "]\n");
 	}
 }
-void print_operator_io_run_flags(int NO, int NI, unsigned int io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI])
+void print_operator_io_run_flags(int NO, int NI, uint64_t io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI])
 {
 	operator_run_opt_flags_t io_run_opts[NO][NI];
 	operator_run_opts_to_io_run_opts(NO, NI, io_flags, io_run_opts, run_opts);
@@ -47,7 +47,7 @@ void print_operator_io_run_flags(int NO, int NI, unsigned int io_flags, operator
 }
 
 
-void operator_run_opts_to_io_run_opts(int NO, int NI, unsigned int io_flags, operator_run_opt_flags_t io_run_opts[NO][NI], operator_run_opt_flags_t run_opts[NO + NI][NO + NI])
+void operator_run_opts_to_io_run_opts(int NO, int NI, uint64_t io_flags, operator_run_opt_flags_t io_run_opts[NO][NI], operator_run_opt_flags_t run_opts[NO + NI][NO + NI])
 {
 
 	int in_index = 0;
@@ -62,7 +62,7 @@ void operator_run_opts_to_io_run_opts(int NO, int NI, unsigned int io_flags, ope
 
 			if (!MD_IS_SET(io_flags, j))
 				continue;
-			
+
 			//printf("NO = %d, NI= %d, out_index = %d, in_index = %d, i = %d, j = %d\n", NO, NI, out_index, in_index, i, j);
 
 			io_run_opts[out_index][in_index] = run_opts[i][j] & run_opts[j][i];
@@ -74,9 +74,9 @@ void operator_run_opts_to_io_run_opts(int NO, int NI, unsigned int io_flags, ope
 	}
 }
 
-void operator_io_run_opts_to_run_opts(int NO, int NI, unsigned int io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], operator_run_opt_flags_t io_run_opts[NO][NI])
+void operator_io_run_opts_to_run_opts(int NO, int NI, uint64_t io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], operator_run_opt_flags_t io_run_opts[NO][NI])
 {
-	
+
 	int in_index = 0;
 	for (int i = 0; i < NO + NI; i++) {
 
@@ -99,19 +99,19 @@ void operator_io_run_opts_to_run_opts(int NO, int NI, unsigned int io_flags, ope
 }
 
 
-unsigned int operator_index_to_io_index(unsigned int io_flags, unsigned int index, bool output)
+unsigned int operator_index_to_io_index(uint64_t io_flags, unsigned int index, bool output)
 {
 	assert(output == (bool)MD_IS_SET(io_flags, index));
 	unsigned int io_index = 0;
-	
+
 	for (unsigned int i = 0; i < index; i++)
 		if (output == (bool)MD_IS_SET(io_flags, i))
 			io_index++;
-	
+
 	return io_index;
 }
 
-unsigned int operator_io_index_to_index(unsigned int io_flags, unsigned int io_index, bool output)
+unsigned int operator_io_index_to_index(uint64_t io_flags, unsigned int io_index, bool output)
 {
 	unsigned int counter = 0;
 	unsigned int index = 0;
@@ -144,27 +144,27 @@ void operator_set_run_opt_sym(int N, operator_run_opt_flags_t run_opts[N][N], un
 	run_opts[j][i] = MD_SET(run_opts[j][i], option);
 }
 
-void operator_set_oi_run_opt(int NO, int NI, unsigned int io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int o, unsigned int i, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
+void operator_set_oi_run_opt(int NO, int NI, uint64_t io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int o, unsigned int i, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
 {
 	unsigned int oind = operator_io_index_to_index(io_flags, o, true);
 	unsigned int iind = operator_io_index_to_index(io_flags, i, false);
-	
+
 	operator_set_run_opt_sym(NO + NI, run_opts, oind, iind, option);
 }
 
-void operator_set_oo_run_opt(int NO, int NI, unsigned int io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int o1, unsigned int o2, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
+void operator_set_oo_run_opt(int NO, int NI, uint64_t io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int o1, unsigned int o2, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
 {
 	unsigned int oind1 = operator_io_index_to_index(io_flags, o1, true);
 	unsigned int oind2 = operator_io_index_to_index(io_flags, o2, true);
-	
+
 	operator_set_run_opt(NO + NI, run_opts, oind1, oind2, option);
 }
 
-void operator_set_ii_run_opt(int NO, int NI, unsigned int io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int i1, unsigned int i2, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
+void operator_set_ii_run_opt(int NO, int NI, uint64_t io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int i1, unsigned int i2, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
 {
 	unsigned int iind1 = operator_io_index_to_index(io_flags, i1, false);
 	unsigned int iind2 = operator_io_index_to_index(io_flags, i2, false);
-	
+
 	operator_set_run_opt(NO + NI, run_opts, iind1, iind2, option);
 }
 
@@ -179,27 +179,27 @@ void operator_unset_run_opt_sym(int N, operator_run_opt_flags_t run_opts[N][N], 
 	run_opts[j][i] = MD_CLEAR(run_opts[j][i], option);
 }
 
-void operator_unset_oi_run_opt(int NO, int NI, unsigned int io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int o, unsigned int i, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
+void operator_unset_oi_run_opt(int NO, int NI, uint64_t io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int o, unsigned int i, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
 {
 	unsigned int oind = operator_io_index_to_index(io_flags, o, true);
 	unsigned int iind = operator_io_index_to_index(io_flags, i, false);
-	
+
 	operator_unset_run_opt_sym(NO + NI, run_opts, oind, iind, option);
 }
 
-void operator_unset_oo_run_opt(int NO, int NI, unsigned int io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int o1, unsigned int o2, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
+void operator_unset_oo_run_opt(int NO, int NI, uint64_t io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int o1, unsigned int o2, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
 {
 	unsigned int oind1 = operator_io_index_to_index(io_flags, o1, true);
 	unsigned int oind2 = operator_io_index_to_index(io_flags, o2, true);
-	
+
 	operator_unset_run_opt(NO + NI, run_opts, oind1, oind2, option);
 }
 
-void operator_unset_ii_run_opt(int NO, int NI, unsigned int io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int i1, unsigned int i2, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
+void operator_unset_ii_run_opt(int NO, int NI, uint64_t io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int i1, unsigned int i2, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
 {
 	unsigned int iind1 = operator_io_index_to_index(io_flags, i1, false);
 	unsigned int iind2 = operator_io_index_to_index(io_flags, i2, false);
-	
+
 	operator_unset_run_opt(NO + NI, run_opts, iind1, iind2, option);
 }
 
@@ -210,17 +210,17 @@ bool operator_get_run_opt(int N, operator_run_opt_flags_t run_opts[N][N], unsign
 
 bool operator_get_run_opt_sym(int N, operator_run_opt_flags_t run_opts[N][N], unsigned int i, unsigned int j, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
 {
-	return MD_IS_SET(run_opts[i][j], option) && MD_IS_SET(run_opts[j][i], option); 
+	return MD_IS_SET(run_opts[i][j], option) && MD_IS_SET(run_opts[j][i], option);
 }
 
-bool operator_get_oi_run_opt(int NO, int NI, unsigned int io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int o, unsigned int i, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
+bool operator_get_oi_run_opt(int NO, int NI, uint64_t io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int o, unsigned int i, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
 {
 	unsigned int oind = operator_io_index_to_index(io_flags, o, true);
 	unsigned int iind = operator_io_index_to_index(io_flags, i, false);
 
 	return operator_get_run_opt_sym(NO + NI, run_opts, oind, iind, option);
 }
-bool operator_get_oo_run_opt(int NO, int NI, unsigned int io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int o1, unsigned int o2, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
+bool operator_get_oo_run_opt(int NO, int NI, uint64_t io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int o1, unsigned int o2, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
 {
 	unsigned int oind1 = operator_io_index_to_index(io_flags, o1, true);
 	unsigned int oind2 = operator_io_index_to_index(io_flags, o2, true);
@@ -228,7 +228,7 @@ bool operator_get_oo_run_opt(int NO, int NI, unsigned int io_flags, operator_run
 	return operator_get_run_opt(NO + NI, run_opts, oind1, oind2, option);
 }
 
-bool operator_get_ii_run_opt(int NO, int NI, unsigned int io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int i1, unsigned int i2, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
+bool operator_get_ii_run_opt(int NO, int NI, uint64_t io_flags, operator_run_opt_flags_t run_opts[NO + NI][NO + NI], unsigned int i1, unsigned int i2, enum OPERATOR_RUN_OPT_FLAGS_INDEX option)
 {
 	unsigned int iind1 = operator_io_index_to_index(io_flags, i1, false);
 	unsigned int iind2 = operator_io_index_to_index(io_flags, i2, false);
@@ -237,7 +237,7 @@ bool operator_get_ii_run_opt(int NO, int NI, unsigned int io_flags, operator_run
 }
 
 
-static void opprops_init(struct opprop_s* n, unsigned int N, unsigned int io_flags, operator_prop_flags_t flags[N][N])
+static void opprops_init(struct opprop_s* n, unsigned int N, uint64_t io_flags, operator_prop_flags_t flags[N][N])
 {
 	n->N = N;
 
@@ -262,14 +262,14 @@ static void opprops_init(struct opprop_s* n, unsigned int N, unsigned int io_fla
 }
 
 
-const struct opprop_s* opprop_create(unsigned int N, unsigned int io_flags, operator_prop_flags_t flags[N][N])
+const struct opprop_s* opprop_create(unsigned int N, uint64_t io_flags, operator_prop_flags_t flags[N][N])
 {
 	PTR_ALLOC(struct opprop_s, n);
 	opprops_init(n, N, io_flags, flags);
 	return PTR_PASS(n);
 }
 
-const struct opprop_s* opprop_io_create(unsigned int NO, unsigned int NI, unsigned int io_flags, operator_prop_flags_t flags[NO][NI])
+const struct opprop_s* opprop_io_create(unsigned int NO, unsigned int NI, uint64_t io_flags, operator_prop_flags_t flags[NO][NI])
 {
 	PTR_ALLOC(struct opprop_s, n);
 
@@ -300,7 +300,7 @@ void opprop_free(const struct opprop_s* x)
 		if (NULL != x->flags)
 			xfree(x->flags);
 		xfree(x);
-	}	
+	}
 }
 
 operator_prop_flags_t opprop_get(const struct opprop_s* x, unsigned int i, unsigned int j)
@@ -353,4 +353,3 @@ void opprop_io_print(int debug_level, const struct opprop_s* x)
 		debug_printf(debug_level, "]\n");
 	}
 }
-
